backend/requirements.txt
fastapi>=0.95.0
fastapi[standard]
uvicorn>=0.22.0
pydantic>=1.10.7
python-dotenv>=1.0.0
openai>=1.0.0
python-multipart>=0.0.6
